% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{calculate_col_width}
\alias{calculate_col_width}
\title{Internal helper function to prepare a data frame and determine the width of its columns for pdf export}
\usage{
calculate_col_width(df, ref)
}
\arguments{
\item{df}{\verb{[data.frame]} A single data frame with named columns.}

\item{ref}{\verb{[character(0+)]} A character vector whose entries specify a selection of columns of \code{df}.
The format of the entries follows: \code{name [label]}.}
}
\value{
The transformed reference column vector whose entries include solely the column names (\code{ref}),
the splitted labels as vectors stored within a list (\code{label_vecs}),
the maximal column widths for all data frame columns (\code{width_max}),
the maximal column widths for the reference columns (\code{ref_width}),
the maximal column widths for the non-reference columns of the dataframe (\code{width}),
the table width exclusively row names (\code{table_width}),
and a logical indicating whether the reference columns specification is valid (\code{check_ref_cols}).
}
\description{
Shortens and splits the column labels if necessary, adapts the entries of the reference column vector for
easier handling, and determines the width of reference columns and data frame columns for upcoming pdf generation.
Widths are given in amount of characters.
}
\keyword{internal}
