% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_receiver}
\alias{check_receiver}
\title{Produce a warning for non-available receiver names}
\usage{
check_receiver(receiver_id, module_ids)
}
\arguments{
\item{receiver_id}{Character string defining the module that should receive a subject identifier
from clinlines.}

\item{module_ids}{Vector of characters defining all available module IDs.}
}
\value{
Logical outcome of the test invisible.
}
\description{
Produce a warning for non-available receiver names
}
\keyword{internal}
