% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{excel_export}
\alias{excel_export}
\title{Internal helper function which performs the download as .xlsx file.}
\usage{
excel_export(data_to_download, file, intended_use_label)
}
\arguments{
\item{data_to_download}{\verb{[list(data.frame)]} A list of data frames to be downloaded.}

\item{file}{\verb{[character(1)]} A string specifying the filename with ending ".xlsx".}
}
\description{
Will export all datasets listed within \code{data_to_download} into a separate
Excel worksheet. Column names of the datasets will be extended by their labels.
}
\keyword{internal}
