% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{export_modal_content}
\alias{export_modal_content}
\title{Internal helper function that specifies the content of the download modal dialogue.}
\usage{
export_modal_content(ns, file_name, cond, colnames, activate_checkbox = FALSE)
}
\arguments{
\item{ns}{A function that expects an ID and returns the namespaced ID, usually generated by \code{shiny::NS()}.}

\item{file_name}{\verb{[character(1)]} A string containing the initial file name.}

\item{cond}{\verb{[character(1)]} A string containing a JavaScript expression that will be evaluated
to determine whether the conditionalPanel should be displayed.}

\item{colnames}{\verb{[character(0+)]} A character vector containing the reference column choices.}

\item{activate_checkbox}{\verb{[logical(1)]} A boolean indicating whether the checkbox should be activated or not.
Default is FALSE.}
}
\value{
A shiny tagList.
}
\description{
Internal helper function that specifies the content of the download modal dialogue.
}
\keyword{internal}
