% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_default_vars}
\alias{fill_default_vars}
\title{Internal helper function to set missing default columns}
\usage{
fill_default_vars(default_vars, dataset)
}
\arguments{
\item{default_vars}{\verb{[list(character(0+))]}

Named list of character vectors which contain the names of columns to be displayed at module
launch per dataset. List entries are named according to the \code{dataset} names but can be NULL for some or all
datasets.}

\item{dataset}{\verb{[list(data.frame)]}

A list of data.framish dataset(s) that will be shown as listings.}
}
\value{
Named list of character vectors which Contain the names of columns to be displayed at module launch for
every dataset. List entries are named according to the \code{dataset} names.
}
\description{
Avoid empty displays in case no default columns were specified by showing the first six columns of a dataset. If the
dataset does contain less than six columns, the whole dataset will be displayed.
}
\keyword{internal}
