% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_choices}
\alias{generate_choices}
\title{Internal helper function to generate either column or dataset choices for the corresponding UI dropdown menues}
\usage{
generate_choices(dataset)
}
\arguments{
\item{dataset}{\verb{[data.frame | list(data.frame)]}

Single data.frame to retrieve its column labels or a list of data.frames to retrieve their name labels. In both
cases the labels will be pasted to the column/dataset names and returned in order to provide meaningful choices in
the corresponding dropdown menues to the end-user.}
}
\value{
Character vector of the type \verb{<name> [<label>]}. Adds "No label" as placeholder for missing labels.
}
\description{
Internal helper function to generate either column or dataset choices for the corresponding UI dropdown menues
}
\keyword{internal}
