% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_labels}
\alias{get_labels}
\title{Internal helper function to gather either column or name labels of data.frames}
\usage{
get_labels(dataset)
}
\arguments{
\item{dataset}{\verb{[data.frame | list(data.frame)]}

Single data.frame to gather its column labels or a list of data.frames to gather their name labels.}
}
\value{
Character vector of labels of \code{dataset}. Named with the respective column or data.frame names. Contains
"No label" for missing labels.
}
\description{
Internal helper function to gather either column or name labels of data.frames
}
\keyword{internal}
