% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_listings.R
\name{listings_UI}
\alias{listings_UI}
\alias{listings_server}
\title{A module that displays datasets as listings}
\usage{
listings_UI(module_id)

listings_server(
  module_id,
  dataset_list,
  default_vars = NULL,
  footers = NULL,
  dataset_metadata,
  pagination = NULL,
  intended_use_label = NULL,
  subjid_var = "USUBJID",
  on_sbj_click = NULL,
  review = NULL,
  exclude_var_names_from_column_headings
)
}
\arguments{
\item{module_id}{Shiny id}

\item{dataset_list}{\verb{[shiny::reactive(list(data.frame)]}

A reactive list of data.framish dataset(s) that will be shown as listing.}

\item{default_vars}{\verb{[list(characters(1+)) | NULL]}

A list of character vectors which contain the variable names to be displayed as default per
dataset. Named according to the \code{dataset_names}. If \code{NULL}, the first six variables are displayed for each
dataset.}

\item{footers}{\verb{[list(characters(1+)) | NULL]}
A list of character vectors that specify per-dataset footer text.
Names should match those provided through \code{dataset_names}.
Each element of the character vector will be displayed on separate lines (see example).}

\item{dataset_metadata}{\verb{[list(character(1), character(1+))]} A list with the following two elements:
\code{dataset_metadata$name()} containing a reactive string specifying the name of the selected
dataset and \code{dataset_metadata$date_range()} containing a reactive character vector with two entries
specifying the earliest and latest modification date in the dataset.
Usually obtained from module manager.}

\item{pagination}{\verb{[logical(1) | NULL]} Either a boolean indicating if pagination should be activated, or
NULL for which pagination will be activated for large datasets (nrows > 1000) automatically.}

\item{intended_use_label}{\verb{[character(1) | NULL]} Either a string indicating the intended use for export, or
NULL. The provided label will be displayed prior to the download and will also be included in the exported file.}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID. Default value is 'USUBJID'}

\item{on_sbj_click}{\verb{[function()]}

Function to invoke when a subject ID is clicked in a listing}

\item{review}{\verb{[list()]}

Configuration of the experimental data review feature.
Only one instance of the listings module can use this feature on any given app.
For more details, please refer to \code{vignette("data_review")}.}

\item{exclude_var_names_from_column_headings}{\verb{[logical(1)]}

Display only dataset variable labels for variables that have them (e.g. turns "VAR_NAME \link{Var Label}" into "Var Label")}
}
\description{
This module displays a given dataset as listing. It allows switching between datasets if it receives
more than one.
}
\section{Functions}{
\itemize{
\item \code{listings_server()}: 

}}
\seealso{
Other data_listings: 
\code{\link{mod_listings}()}
}
\concept{data_listings}
