% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_listings.R
\name{mock_listings_app}
\alias{mock_listings_app}
\title{Mock app without layout}
\usage{
mock_listings_app(mode = "single")
}
\arguments{
\item{mode}{Character value that indicates whether to include one or multiple datasets. Could either be "single" or
"multi".}
}
\description{
\code{mock_listings_app} launches a mock app for the dv.listings shiny module. Only for development purposes.
}
\seealso{
Other mock apps: 
\code{\link{mock_listings_mm}()}
}
\concept{mock apps}
\keyword{internal}
