% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_listings.R
\name{mock_listings_mm}
\alias{mock_listings_mm}
\title{Mock app integrated in the module manager}
\usage{
mock_listings_mm()
}
\description{
\code{mock_listings_mm} launches a mock app for the dv.listings shiny module by means of
the module manager (dv.manager).
}
\seealso{
Other mock apps: 
\code{\link{mock_listings_app}()}
}
\concept{mock apps}
\keyword{internal}
