% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_listings.R
\name{mod_export_listings_UI}
\alias{mod_export_listings_UI}
\title{Create user interface for the export listings shiny module of \pkg{dv.listings}}
\usage{
mod_export_listings_UI(module_id)
}
\arguments{
\item{module_id}{\verb{[character(1)]} A unique ID string to create a namespace. Must match the ID of
\code{mod_exp_listings_server()}.}
}
\value{
A shiny \code{uiOutput} element.
}
\description{
Create user interface for the export listings shiny module of \pkg{dv.listings}
}
\keyword{internal}
