% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{pdf_export}
\alias{pdf_export}
\title{Internal helper function which performs the download as .pdf file.}
\usage{
pdf_export(
  data_to_download,
  ref_cols,
  file,
  metadata,
  active_session = TRUE,
  intended_use_label,
  keep_tex = FALSE
)
}
\arguments{
\item{data_to_download}{\verb{[list(data.frame)]} A named list containing one data frame to be downloaded.
#' @param ref \verb{[character(0+)]} A character vector whose entries specify a selection of columns of the
dataset to download. The format of the entries follows: \code{name [label]}.}

\item{file}{\verb{[character(1)]} A string specifying the filename with ending ".pdf".}

\item{metadata}{\verb{[character(3)]} A character vector specifying the dataset's name, date, and an
additional footnote text.}

\item{active_session}{\verb{[logical(1)]} Logical value that indicates if the helper function is used within a shiny
session. If yes, the parts for displaying a progress bar get activated. Defaults to \code{TRUE}.}
}
\value{
Number of PDF pages that are generated.
}
\description{
Will export the dataset listed within \code{data_to_download} into a PDF file.
}
\keyword{internal}
