% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{pdf_preprocessing}
\alias{pdf_preprocessing}
\title{Internal helper function to divide a data frame into smaller data frames that fit on one PDF page each}
\usage{
pdf_preprocessing(df, ref)
}
\arguments{
\item{df}{\verb{[data.frame]} A single data frame with named columns.}

\item{ref}{\verb{[character(0+)]} A character vector whose entries specify a selection of columns of \code{df}.
The format of the entries follows: \code{name [label]}.}
}
\value{
Named list containing a list of data frames (\code{list_of_df}) whose entries fit on one PDF page each.
}
\description{
Needed as preprocessing step before PDF creation by means of RMarkdown can take place.
}
\keyword{internal}
