% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_data}
\alias{set_data}
\title{Prepare data as it should be displayed in the module}
\usage{
set_data(base_data, selector)
}
\arguments{
\item{base_data}{\verb{[data.frame]}

Single data.frame.}

\item{selector}{\verb{[character(0+)]}

Character vector of a selection of column names from \code{base_data}.}
}
\value{
A list of tree elements: \code{data} is the data.frame prepared to be displayed as-is. \code{col_names} is a
vector of column names created by combining the variable names with their labels.
\code{row_names} is a character vector containing the number of the corresponding row.
}
\description{
Reduces the dataset to only those columns specified in \code{selector} and produces descriptive column names
by adding their labels to the variable names.
}
\keyword{internal}
