% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_labels}
\alias{set_labels}
\title{Internal helper function to set labels for all dataset columns}
\usage{
set_labels(dataset, labels)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

Single data.frame}

\item{labels}{Character vector of labels named with the respective column names \code{dataset}.}
}
\value{
A data.frame with column labels
}
\description{
Internal helper function to set labels for all dataset columns
}
\keyword{internal}
