% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{shorten_entries}
\alias{shorten_entries}
\title{Internal helper function to cut strings that exceed a specific length}
\usage{
shorten_entries(vec, len_max)
}
\arguments{
\item{vec}{\verb{[character(0+)]} A character vector containing the strings to be cut if exceeding a maximal length.}

\item{len_max}{\verb{[integer(1)]} An integer specifying the maximal length, must be equal to or greater than 3.}
}
\value{
A character vector containing the shortened strings in case they exceeded the maximal length.
}
\description{
Shortens strings that are too long and denotes cutting by pasting "..." at the end of the string.
}
\keyword{internal}
