% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{split_label}
\alias{split_label}
\title{Internal helper function to prepare a data frame and determine the width of its columns for pdf export}
\usage{
split_label(label, min_width, max_width, label_width)
}
\arguments{
\item{label}{\verb{[character(1)]} A string representing the label.}

\item{min_width}{\verb{[integer(1)]} An integer specifying the minimal width of the column (not less than 1).}

\item{max_width}{\verb{[integer(1)]} An integer specifying the maximal width of the column
(greater than 3 and also not less than min_width \verb{[integer(1)]}).}

\item{label_width}{\verb{[integer(1)]} An integer specifying the maximal number of rows which is allocated for the label
(not less than 1).}
}
\value{
A list containing the \code{label_vec} which includes the splitted label and the actual width \code{col_width}.
}
\description{
Shortens and splits the column label if necessary, so that it fits into \code{label_width} rows of its column.
}
\keyword{internal}
