% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{warn_function}
\alias{warn_function}
\title{Internal helper function for warnings in case of missing/wrong selections done by the user.}
\usage{
warn_function(cond, input_id, text)
}
\arguments{
\item{cond}{\verb{[logical(1)]} Logical value that indicates whether the warning should be triggered (if \code{TRUE}) or
not (if \code{FALSE}).}

\item{input_id}{\verb{[character(1)]} ID string indicating at which input field the warning should appear.}

\item{text}{\verb{[character(1)]} A string containing the warning message to be displayed.}
}
\description{
Triggers a warning feedback if a particular condition is true and hides the
feedback if it is false by using \pkg{shinyFeedback}.
}
\keyword{internal}
