% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{check_review_parameter}
\alias{check_review_parameter}
\title{Early error feedback function for the optional review parameter}
\usage{
check_review_parameter(datasets, dataset_names, review, err)
}
\arguments{
\item{datasets}{\verb{[list(data.frame)]}

Available datasets for review.}

\item{dataset_names}{\verb{[character(n)]}

Names of the datasets provided by the previous parameter.}

\item{review}{\verb{[list()]}

Configuration of the experimental data review feature. Please refer to \code{vignette("data_review")}.}

\item{error}{\verb{[environment]}
This environment has at least one element named "messages". It is a character vector. Diagnostic messages related to
the configuration of the review parameter will be placed here.}
}
\description{
Early error feedback function for the optional review parameter
}
