% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_listings.R
\name{mod_export_listings_server}
\alias{mod_export_listings_server}
\title{Create server for export listings shiny module of \pkg{dv.listings}}
\usage{
mod_export_listings_server(
  module_id,
  dataset_metadata,
  dataset_list,
  data,
  data_selection_name,
  current_rows,
  intended_use_label,
  footers
)
}
\arguments{
\item{module_id}{\verb{[character(1)]} A unique ID string to create a namespace. Must match the ID of
\code{mod_exp_listings_UI()}.}

\item{dataset_metadata}{\verb{[list(character(1), character(1+))]} A list with the following two elements:
\code{dataset_metadata$name()} containing a reactive string specifying the name of the selected
dataset and \code{dataset_metadata$date_range()} containing a reactive character vector with two entries
specifying the earliest and latest modification date in the dataset.
Usually obtained from module manager.}

\item{dataset_list}{\verb{[shiny::reactive(list(data.frame))]} A reactive list of named datasets.
Usually obtained from module manager.}

\item{data}{\verb{[shiny::reactive(list(data.frame, character(0+)))]} A reactive list with the
following two elements: \code{data()$data} containing a single data frame and \code{data()$col_names}
containing the column names of \code{data()$data}.}

\item{data_selection_name}{\verb{[shiny::reactive(character(1))]} A reactive string specifying the name of
\code{data()$data}.}

\item{current_rows}{\verb{[shiny::reactive(character(0+))]} A reactive integer vector containing a selection
of row numbers from \code{data()$data}.}

\item{intended_use_label}{\verb{[character(1) | NULL]} Either a character indicating the intended use for the download, or
NULL. If a label is provided it will be shown before the download and will also be included in the downloaded file.}

\item{footers}{\verb{[list(character(1+)) | NULL]} Pass-through argument from \code{mod_listings} specifying per-dataset footer
text.}
}
\description{
Create server for export listings shiny module of \pkg{dv.listings}
}
\keyword{internal}
