% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{prep_export_data}
\alias{prep_export_data}
\title{Internal helper function for preparing the dataset(s) to download.}
\usage{
prep_export_data(
  data_selection,
  current_data,
  data_selection_name,
  dataset_list
)
}
\arguments{
\item{data_selection}{\verb{[character(1)]} Either \code{"single"} or \code{"all"} depending on whether the
currently displayed dataset (\code{current_data}) or all datasets (\code{dataset_list}) should be downloaded.}

\item{current_data}{\verb{[data.frame]} A single data frame with named columns.}

\item{data_selection_name}{\verb{[character(1)]} A string specifying the name of \code{current_data}.}

\item{dataset_list}{\verb{[list(data.frame)]} A list of named datasets.}
}
\value{
Named list containing the data frames which are now ready for download.
}
\description{
Internal helper function for preparing the dataset(s) to download.
}
\keyword{internal}
