% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_up_datatable}
\alias{set_up_datatable}
\title{Internal helper function to set up colnames, rownames, and paging arguments for DT's datatable.
Main purpose is to easily test returned arguments.}
\usage{
set_up_datatable(dataset, pagination, exclude_var_names_from_column_headings)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

Single data.frame}

\item{pagination}{\verb{[logical(1) | NULL]}

Either a boolean indicating if pagination should be activated, or
NULL for which pagination will be activated for large datasets (nrows > 1000) automatically.}

\item{exclude_var_names_from_column_headings}{\verb{[logical(1)]}

Use only dataset variable labels for variables that have them (output "Var Label" instead of "VAR_NAME [Var Label]").}
}
\value{
List containing character vectors for column names and row names and
a logical value for de-/activating paging
}
\description{
Internal helper function to set up colnames, rownames, and paging arguments for DT's datatable.
Main purpose is to easily test returned arguments.
}
\keyword{internal}
